initial 2.3.9 release commit

+++ docker/util/remove_down_nodes.sh
@@ -1,4 +1,6 @@
 #!/usr/bin/env bash
 
-sudo docker node ls --filter role=worker|grep Down|grep -o '^\S\+'|xargs sudo docker node rm
-
+export DOWN_NODES_NUM=$(sudo docker node ls --filter role=worker|grep Down|grep -o '^\S\+'|wc -l)
+if [ "${DOWN_NODES_NUM}" -gt 0 ]; then
+  sudo docker node ls --filter role=worker|grep Down|grep -o '^\S\+'|xargs sudo docker node rm;
+fi
